+++ mkosi/__init__.py
@@ -23,6 +23,7 @@ from typing import Optional, TextIO, Union, cast
 
 import mkosi.resources
 from mkosi.archive import extract_tar, make_cpio, make_tar
+from mkosi.bubblewrap import bwrap, chroot_cmd
 from mkosi.burn import run_burn
 from mkosi.config import (
     BiosBootloader,
@@ -50,14 +51,12 @@ from mkosi.installer import clean_package_manager_metadata, package_manager_scri
 from mkosi.kmod import gen_required_kernel_modules, process_kernel_modules
 from mkosi.log import ARG_DEBUG, complete_step, die, log_notice, log_step
 from mkosi.manifest import Manifest
-from mkosi.mounts import mount, mount_overlay, mount_usr
+from mkosi.mounts import mount_overlay, mount_usr
 from mkosi.pager import page
 from mkosi.partition import Partition, finalize_root, finalize_roothash
 from mkosi.qemu import KernelType, QemuDeviceNode, copy_ephemeral, run_qemu, run_ssh
 from mkosi.run import (
     become_root,
-    bwrap,
-    chroot_cmd,
     find_binary,
     fork_and_wait,
     init_mount_namespace,
@@ -65,7 +64,7 @@ from mkosi.run import (
     run_openssl,
 )
 from mkosi.state import MkosiState
-from mkosi.tree import copy_tree, install_tree, move_tree, rmtree
+from mkosi.tree import copy_tree, move_tree, rmtree
 from mkosi.types import PathString
 from mkosi.util import (
     INVOKING_USER,
@@ -107,7 +106,7 @@ def mount_base_trees(state: MkosiState) -> Iterator[None]:
             if path.is_dir():
                 bases += [path]
             elif path.suffix == ".tar":
-                extract_tar(path, d)
+                extract_tar(state, path, d)
                 bases += [d]
             elif path.suffix == ".raw":
                 run(["systemd-dissect", "-M", path, d])
@@ -354,23 +353,15 @@ def mount_build_overlay(state: MkosiState, volatile: bool = False) -> Iterator[P
 
 
 @contextlib.contextmanager
-def finalize_mounts(config: MkosiConfig) -> Iterator[list[PathString]]:
+def finalize_source_mounts(config: MkosiConfig) -> Iterator[list[PathString]]:
     with contextlib.ExitStack() as stack:
-        sources = [
+        mounts = [
             (stack.enter_context(mount_overlay([source])) if config.build_sources_ephemeral else source, target)
             for source, target
             in [(Path.cwd(), Path.cwd())] + [t.with_prefix(Path.cwd()) for t in config.build_sources]
         ]
 
-        # bwrap() mounts /home and /var read-only during execution. So let's add the bind mount options for the
-        # directories that could be in /home or /var that we do need to be writable.
-        sources += [
-            (d, d)
-            for d in (config.workspace_dir_or_default(), config.cache_dir, config.output_dir, config.build_dir)
-            if d
-        ]
-
-        yield flatten(["--bind", src, target] for src, target in sorted(set(sources), key=lambda s: s[1]))
+        yield flatten(["--bind", src, target] for src, target in sorted(set(mounts), key=lambda s: s[1]))
 
 
 def script_maybe_chroot(script: Path, mountpoint: str) -> list[str]:
@@ -451,7 +442,7 @@ def run_prepare_scripts(state: MkosiState, build: bool) -> None:
             step_msg = "Running prepare script {}…"
             arg = "final"
 
-        mounts = stack.enter_context(finalize_mounts(state.config))
+        sources = stack.enter_context(finalize_source_mounts(state.config))
         cd = stack.enter_context(finalize_chroot_scripts(state))
 
         for script in state.config.prepare_scripts:
@@ -475,10 +466,10 @@ def run_prepare_scripts(state: MkosiState, build: bool) -> None:
 
             with complete_step(step_msg.format(script)):
                 bwrap(
+                    state,
                     script_maybe_chroot(script, "/work/prepare") + [arg],
                     network=True,
-                    readonly=True,
-                    options=mounts,
+                    options=sources,
                     scripts=hd,
                     env=env | state.config.environment,
                     stdin=sys.stdin,
@@ -516,7 +507,7 @@ def run_build_scripts(state: MkosiState) -> None:
     with (
         mount_build_overlay(state, volatile=True),
         finalize_chroot_scripts(state) as cd,
-        finalize_mounts(state.config) as mounts,
+        finalize_source_mounts(state.config) as sources,
     ):
         for script in state.config.build_scripts:
             helpers = {
@@ -548,10 +539,10 @@ def run_build_scripts(state: MkosiState) -> None:
                 complete_step(f"Running build script {script}…"),
             ):
                 bwrap(
+                    state,
                     script_maybe_chroot(script, "/work/build-script") + cmdline,
                     network=state.config.with_network,
-                    readonly=True,
-                    options=mounts,
+                    options=sources,
                     scripts=hd,
                     env=env | state.config.environment,
                     stdin=sys.stdin,
@@ -577,7 +568,7 @@ def run_postinst_scripts(state: MkosiState) -> None:
 
     with (
         finalize_chroot_scripts(state) as cd,
-        finalize_mounts(state.config) as mounts,
+        finalize_source_mounts(state.config) as sources,
     ):
         for script in state.config.postinst_scripts:
             helpers = {
@@ -603,10 +594,10 @@ def run_postinst_scripts(state: MkosiState) -> None:
                 complete_step(f"Running postinstall script {script}…"),
             ):
                 bwrap(
+                    state,
                     script_maybe_chroot(script, "/work/postinst") + ["final"],
                     network=state.config.with_network,
-                    readonly=True,
-                    options=mounts,
+                    options=sources,
                     scripts=hd,
                     env=env | state.config.environment,
                     stdin=sys.stdin,
@@ -632,7 +623,7 @@ def run_finalize_scripts(state: MkosiState) -> None:
 
     with (
         finalize_chroot_scripts(state) as cd,
-        finalize_mounts(state.config) as mounts,
+        finalize_source_mounts(state.config) as sources,
     ):
         for script in state.config.finalize_scripts:
             helpers = {
@@ -658,10 +649,10 @@ def run_finalize_scripts(state: MkosiState) -> None:
                 complete_step(f"Running finalize script {script}…"),
             ):
                 bwrap(
+                    state,
                     script_maybe_chroot(script, "/work/finalize"),
                     network=state.config.with_network,
-                    readonly=True,
-                    options=mounts,
+                    options=sources,
                     scripts=hd,
                     env=env | state.config.environment,
                     stdin=sys.stdin,
@@ -1152,6 +1143,7 @@ def prepare_grub_bios(state: MkosiState, partitions: Sequence[Partition]) -> Non
         earlyconfig.flush()
 
         bwrap(
+            state,
             [
                 mkimage,
                 "--directory", directory,
@@ -1165,7 +1157,7 @@ def prepare_grub_bios(state: MkosiState, partitions: Sequence[Partition]) -> Non
                 "part_gpt",
                 "biosdisk",
                 "search",
-                "search_fs_file"
+                "search_fs_file",
             ],
             options=["--bind", state.root / "usr", "/usr"],
         )
@@ -1210,12 +1202,41 @@ def install_grub_bios(state: MkosiState, partitions: Sequence[Partition]) -> Non
     with complete_step("Installing grub boot loader…"):
         # We don't setup the mountinfo bind mount with bwrap because we need to know the child process pid to
         # be able to do the mount and we don't know the pid beforehand.
-        bwrap(["sh", "-c", f"mount --bind {mountinfo} /proc/$$/mountinfo && exec $0 \"$@\"",
-               setup,
-               "--directory", state.root / "efi" / prefix / "i386-pc",
-               *(["--verbose"] if ARG_DEBUG.get() else []),
-               state.staging / state.config.output_with_format],
-              options=["--bind", state.root / "usr", "/usr"])
+        bwrap(
+            state,
+            [
+                "sh", "-c", f"mount --bind {mountinfo} /proc/$$/mountinfo && exec $0 \"$@\"",
+                setup,
+                "--directory", state.root / "efi" / prefix / "i386-pc",
+                *(["--verbose"] if ARG_DEBUG.get() else []),
+                state.staging / state.config.output_with_format,
+            ],
+            options=["--bind", state.root / "usr", "/usr"],
+        )
+
+
+def install_tree(
+    state: MkosiState,
+    src: Path,
+    dst: Path,
+    target: Optional[Path] = None,
+) -> None:
+    t = dst
+    if target:
+        t = dst / target.relative_to("/")
+
+    with umask(~0o755):
+        t.parent.mkdir(parents=True, exist_ok=True)
+
+    if src.is_dir() or (src.is_file() and target):
+        copy_tree(src, t, preserve_owner=False, use_subvolumes=state.config.use_subvolumes)
+    elif src.suffix == ".tar":
+        extract_tar(state, src, t)
+    elif src.suffix == ".raw":
+        run(["systemd-dissect", "--copy-from", src, "/", t])
+    else:
+        # If we get an unknown file without a target, we just copy it into /.
+        copy_tree(src, t, preserve_owner=False, use_subvolumes=state.config.use_subvolumes)
 
 
 def install_base_trees(state: MkosiState) -> None:
@@ -1224,7 +1245,7 @@ def install_base_trees(state: MkosiState) -> None:
 
     with complete_step("Copying in base trees…"):
         for path in state.config.base_trees:
-            install_tree(path, state.root, use_subvolumes=state.config.use_subvolumes)
+            install_tree(state, path, state.root)
 
 
 def install_skeleton_trees(state: MkosiState) -> None:
@@ -1233,7 +1254,7 @@ def install_skeleton_trees(state: MkosiState) -> None:
 
     with complete_step("Copying in skeleton file trees…"):
         for tree in state.config.skeleton_trees:
-            install_tree(tree.source, state.root, tree.target, use_subvolumes=state.config.use_subvolumes)
+            install_tree(state, tree.source, state.root, tree.target)
 
 
 def install_package_manager_trees(state: MkosiState) -> None:
@@ -1242,12 +1263,7 @@ def install_package_manager_trees(state: MkosiState) -> None:
 
     with complete_step("Copying in package manager file trees…"):
         for tree in state.config.package_manager_trees:
-            install_tree(
-                tree.source,
-                state.workspace / "pkgmngr",
-                tree.target,
-                use_subvolumes=state.config.use_subvolumes
-            )
+            install_tree(state, tree.source, state.workspace / "pkgmngr", tree.target)
 
 
 def install_extra_trees(state: MkosiState) -> None:
@@ -1256,7 +1272,7 @@ def install_extra_trees(state: MkosiState) -> None:
 
     with complete_step("Copying in extra file trees…"):
         for tree in state.config.extra_trees:
-            install_tree(tree.source, state.root, tree.target, use_subvolumes=state.config.use_subvolumes)
+            install_tree(state, tree.source, state.root, tree.target)
 
 
 def install_build_dest(state: MkosiState) -> None:
@@ -1339,6 +1355,7 @@ def build_initrd(state: MkosiState) -> Path:
         *(["--hostname", state.config.hostname] if state.config.hostname else []),
         *(["--root-password", rootpwopt] if rootpwopt else []),
         *([f"--environment={k}='{v}'" for k, v in state.config.environment.items()]),
+        *(["--tools-tree", str(state.config.tools_tree)] if state.config.tools_tree else []),
         *(["-f"] * state.args.force),
     ]
 
@@ -1391,7 +1408,7 @@ def build_microcode_initrd(state: MkosiState) -> Optional[Path]:
             for p in intel.iterdir():
                 f.write(p.read_bytes())
 
-    make_cpio(root, microcode)
+    make_cpio(state, root, microcode)
 
     return microcode
 
@@ -1402,7 +1419,7 @@ def build_kernel_modules_initrd(state: MkosiState, kver: str) -> Path:
         return kmods
 
     make_cpio(
-        state.root, kmods,
+        state, state.root, kmods,
         gen_required_kernel_modules(
             state.root, kver,
             state.config.kernel_modules_initrd_include,
@@ -1669,7 +1686,7 @@ def install_uki(state: MkosiState, partitions: Sequence[Partition]) -> None:
 
 def make_uki(state: MkosiState, stub: Path, kver: str, kimg: Path, output: Path) -> None:
     microcode = build_microcode_initrd(state)
-    make_cpio(state.root, state.workspace / "initrd")
+    make_cpio(state, state.root, state.workspace / "initrd")
     maybe_compress(state.config, state.config.compress_output, state.workspace / "initrd", state.workspace / "initrd")
 
     initrds = [microcode] if microcode else []
@@ -2619,9 +2636,9 @@ def build_image(args: MkosiArgs, config: MkosiConfig) -> None:
         copy_initrd(state)
 
         if state.config.output_format == OutputFormat.tar:
-            make_tar(state.root, state.staging / state.config.output_with_format)
+            make_tar(state, state.root, state.staging / state.config.output_with_format)
         elif state.config.output_format == OutputFormat.cpio:
-            make_cpio(state.root, state.staging / state.config.output_with_format)
+            make_cpio(state, state.root, state.staging / state.config.output_with_format)
         elif state.config.output_format == OutputFormat.uki:
             assert stub and kver and kimg
             make_uki(state, stub, kver, kimg, state.staging / state.config.output_with_format)
@@ -3016,72 +3033,6 @@ def finalize_tools(args: MkosiArgs, images: Sequence[MkosiConfig]) -> Sequence[M
     return new
 
 
-@contextlib.contextmanager
-def mount_tools(tree: Optional[Path]) -> Iterator[None]:
-    if not tree:
-        yield
-        return
-
-    with contextlib.ExitStack() as stack:
-        stack.enter_context(mount_usr(tree))
-
-        # On top of /usr, we also need the certificates, keyrings and crypto policies from the tools tree which are
-        # unfortunately not always in /usr. To make things work, we mount those directories into the host as well.
-        # Because the directories might not exist on the host (which means we can't mount over them), we mount a
-        # writable directory on top of their parent directory using overlayfs so we can create the required mountpoints
-        # without running into permission errors.
-
-        overlays = set()
-        dirs = [
-            subdir
-            for subdir in (
-                Path("etc/pki"),
-                Path("etc/ssl"),
-                Path("etc/crypto-policies"),
-                Path("etc/ca-certificates"),
-                Path("etc/pacman.d"),
-                Path("var/lib/ca-certificates"),
-            )
-            if (tree / subdir).exists()
-        ]
-
-        for subdir in dirs:
-            if (Path("/") / subdir).exists():
-                continue
-
-            if subdir.parent not in overlays:
-                tmp = stack.enter_context(tempfile.TemporaryDirectory(dir="/var/tmp"))
-                stack.enter_context(mount_overlay([Path("/") / subdir.parent], Path(tmp), Path("/") / subdir.parent))
-                overlays.add(subdir.parent)
-
-            (Path("/") / subdir).mkdir(parents=True, exist_ok=True)
-
-        for subdir in dirs:
-            stack.enter_context(
-                mount(what=tree / subdir, where=Path("/") / subdir, operation="--bind", read_only=True)
-            )
-
-        yield
-
-
-@contextlib.contextmanager
-def hide_host_directories() -> Iterator[None]:
-    with contextlib.ExitStack() as stack:
-
-        # We want to limit the effect of host specific admin configuration on image builds, so we hide various config
-        # directories that we can't override using CLI options or environment variables so that they don't affect our
-        # builds.
-
-        for d in ("/etc/rpm",):
-            if not Path(d).exists():
-                continue
-
-            tmp = stack.enter_context(tempfile.TemporaryDirectory())
-            stack.enter_context(mount(what=tmp, where=Path(d), operation="--bind"))
-
-        yield
-
-
 def check_workspace_directory(config: MkosiConfig) -> None:
     wd = config.workspace_dir_or_default()
 
@@ -3214,8 +3165,7 @@ def run_verb(args: MkosiArgs, images: Sequence[MkosiConfig]) -> None:
 
             with (
                 complete_step(f"Building {config.name()} image"),
-                mount_tools(config.tools_tree),
-                hide_host_directories(),
+                mount_usr(config.tools_tree),
                 prepend_to_environ_path(config),
             ):
                 # After tools have been mounted, check if we have what we need
